Onnx backend infer call initial work.

+++ tiny_inference_server/server_core/backend/include/infer_request.hpp
@@ -8,7 +8,7 @@ namespace tie::backend
 struct infer_request
 {
     std::string_view model_name;
-    std::vector<float> data;
+    std::vector<uint8_t> data;
     std::vector<int64_t> shape;
 };
